align file names and class per file structure with Swift guidelines
