+++ mac/keyrace-mac/KeyTap.swift
@@ -1,30 +1,8 @@
-//
-//  keyrace-mac
-//
-//  Created by Nat Friedman on 1/2/21.
-//
+// Created by Thomas Dohmke on 2/19/21.
 
-// TODO:
-// - Restructure, cleanup, learn swift
-// - Save username locally along with token
-// - Talk to server and show leaderboard
-
-
-import SwiftUI
 import Foundation
 import Accessibility
 import Cocoa
-
-@available(OSX 11.0, *)
-@main
-struct MenuBarPopoverApp: App {
-    @NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
-    var body: some Scene {
-        Settings{
-            EmptyView()
-        }
-    }
-}
 
 func formatCount(count: Int) -> String {
     var str = ""
@@ -47,7 +25,7 @@ func formatCount(count: Int) -> String {
         default:
             pfx = ""
         }
-        
+
         var sfx = ""
         if (count < 100) {
             sfx = " today"
@@ -66,7 +44,7 @@ func myCGEventCallback(proxy: CGEventTapProxy, type: CGEventType, event: CGEvent
         var length = 0
         event.keyboardGetUnicodeString(maxStringLength: 1, actualStringLength: &length, unicodeString: &char)
         keyTap.increment(char)
-        
+
         keyTap.appDelegate.menubarItem?.statusBarItem.button?.title = formatCount(count: keyTap.keycount)
     }
 
@@ -94,31 +72,31 @@ class KeyTap {
     var minutes = [Int](repeating:0, count:1440)
     var keys = [Int](repeating: 0, count:256)
     var leaderboardText = NSMutableAttributedString()
-    
+
     init(_ appd: AppDelegate) {
         self.appDelegate = appd
     }
-    
+
     func increment(_ keyCode: UInt16) {
         let date = Date()
         let calendar = Calendar.current
-        
+
         // Reset to 0 at midnight
         let day = calendar.component(.day, from:date)
         if (lastDay != day) {
             lastDay = day
             keycount = 0
             keys = [Int](repeating:0, count:256)
-            
+
             //  Clears our minutes, leaving the last 20 minutes, and than deletes the rest after 20 minutes
             minutes.replaceSubrange(0..<1420, with: repeatElement(0, count: 1420))
             DispatchQueue.main.asyncAfter(deadline: .now() + .seconds(1200), execute: {
                 self.minutes.replaceSubrange(1420..<1440, with: repeatElement(0, count: 20))
             })
         }
-        
+
         keycount += 1
-        
+
         let hour = calendar.component(.hour, from: date)
         let minute = calendar.component(.minute, from: date)
         minutes[hour*60 + minute] += 1
@@ -134,10 +112,10 @@ class KeyTap {
             lastMin = minute
             uploadCount()
         }
-        
+
         saveCount()
     }
-    
+
     func getMinutesChart() -> [Int] {
         // Return the last 20 minutes minutely
         let date = Date()
@@ -145,7 +123,7 @@ class KeyTap {
         let hour = calendar.component(.hour, from: date)
         let min = calendar.component(.minute, from: date)
         let currMin = hour*60 + min
-        
+
         var mins : [Int] = []
         for i in (0...20).reversed() {
             currMin - i > 0 ? mins.append(minutes[currMin - i]): mins.append(minutes[1440 + i - (20 - currMin)])
@@ -159,7 +137,7 @@ class KeyTap {
         for i in 0..<minutes.count {
             hours[i/60] += minutes[i]
         }
-        
+
         return hours
     }
 
@@ -176,7 +154,7 @@ class KeyTap {
     func getLeaderboardText() -> NSMutableAttributedString {
         return leaderboardText
     }
-    
+
     func uploadKeycount() {
         if appDelegate.gh?.token == nil {
             return
@@ -193,7 +171,7 @@ class KeyTap {
 
         var request = URLRequest(url: url.url!)
         request.addValue("Bearer \(appDelegate.gh!.token!)", forHTTPHeaderField: "Authorization")
-        
+
         let task = URLSession.shared.dataTask(with: request) { data, response, error in
             guard let data = data,                            // is there data
                 let response = response as? HTTPURLResponse,  // is there HTTP response
@@ -208,7 +186,7 @@ class KeyTap {
         }
         task.resume()
     }
-    
+
     func parseJSON(json: Data) {
         let decoder = JSONDecoder()
 
@@ -217,7 +195,7 @@ class KeyTap {
             self.leaderboardText = NSMutableAttributedString()
             let attrBlankLine = NSMutableAttributedString(string: " \n")
             self.leaderboardText.append(attrBlankLine)
-            
+
             // Add paragraph styling
             let paragraphStyle = NSMutableParagraphStyle()
             paragraphStyle.lineSpacing = 8
@@ -226,7 +204,7 @@ class KeyTap {
             // Add the tab stops so things are well aligned.
             paragraphStyle.tabStops = [NSTextTab(textAlignment: NSTextAlignment.left, location: 150, options: [:])]
             paragraphStyle.headIndent = 150
-            
+
             for (i, player) in leaderboard.enumerated(){
                 let fullUsername = "    @" + player.username
                 var score = String(format: "\t%d", player.score)
@@ -236,7 +214,7 @@ class KeyTap {
                 }
                 // Add the new line.
                 score += "\n"
-                
+
                 // Create the image for the avatar.
                 var attrImage = NSMutableAttributedString()
                 DispatchQueue.main.sync {
@@ -255,7 +233,7 @@ class KeyTap {
                                                   range: .init(location: 0, length: 1))
                     }
                 }
-                
+
                 // Do the font styling for the line.
                 let attrLine = NSMutableAttributedString(string: fullUsername + score)
                 attrLine.addAttribute(.font, value: NSFont.monospacedSystemFont(ofSize: 13, weight: .regular), range: NSRange(location: 0, length: fullUsername.count + score.count))
@@ -263,23 +241,23 @@ class KeyTap {
                 attrLine.addAttribute(.link,
                                       value: NSURL(string: "https://github.com/"+player.username)!,
                                           range: usernameRange)
-                
+
                 // Add the image.
                 attrLine.replaceCharacters(in: NSRange(location: 0, length: 2), with: attrImage)
                 self.leaderboardText.append(attrLine)
             }
-            
+
             // Set the paragraph styling.
             self.leaderboardText.addAttribute(.paragraphStyle, value: paragraphStyle, range: NSRange(location: 0, length: self.leaderboardText.length))
             self.leaderboardText.setAlignment(.justified, range: NSRange(location: 0, length: self.leaderboardText.length))
         }
     }
-    
+
     func setupKeyTap() {
         if (keyTrapSetup) {
             return
         }
-        
+
         loadCount()
 
         let eventMask = (1 << CGEventType.keyDown.rawValue)
@@ -300,14 +278,14 @@ class KeyTap {
         CGEvent.tapEnable(tap: eventTap, enable: true)
 
         appDelegate.menubarItem!.statusBarItem.button?.title = formatCount(count: keycount)
-        
+
         uploadCount()
     }
-    
+
     func uploadCount () {
         uploadKeycount()
     }
-    
+
     func saveCount() {
         var filename = FileManager.default.homeDirectoryForCurrentUser.appendingPathComponent(".keyrace.tmp")
 
@@ -318,7 +296,7 @@ class KeyTap {
             NSLog("Could not write keycount to \(filename.path)")
             // failed to write file – bad permissions, bad filename, missing permissions, or more likely it can't be converted to the encoding
         }
-        
+
         filename = FileManager.default.homeDirectoryForCurrentUser.appendingPathComponent(".keyrace.minutes.tmp")
         let minStrings = minutes.map({ String($0) })
         str = minStrings.joined(separator: ",")
@@ -340,7 +318,7 @@ class KeyTap {
         }
 
     }
-    
+
     func loadCount() {
         let date = Date()
         let calendar = Calendar.current
@@ -354,13 +332,13 @@ class KeyTap {
             let mtime = attr[FileAttributeKey.modificationDate] as! Date
             lastDay = calendar.component(.day, from:mtime)
             let lastMonth = calendar.component(.month, from:mtime)
-            
+
             if (lastDay == day && lastMonth == month) {
                 let str = try String(contentsOf: filename, encoding: .utf8)
                 keycount = Int(str) ?? 0
             }
         } catch { }
-        
+
         // Load the hourly histogram
         filename = FileManager.default.homeDirectoryForCurrentUser.appendingPathComponent(".keyrace.minutes.tmp")
         do {
@@ -401,7 +379,7 @@ class KeyTap {
             self.setupKeyTap()
             return
         }
-        
+
         // Wait for the user to give us permission
         if (timerRunning) { return }
         self.timerRunning = true
@@ -414,43 +392,5 @@ class KeyTap {
                 self.timerRunning = false
             }
         }
-    }
-}
-
-class AppDelegate: NSObject, NSApplicationDelegate {
-    var statusBarItem: NSStatusItem!
-    var menubarItem: MenubarItem?
-    var keyTap : KeyTap?
-    var gh : GitHub?
-    
-    func applicationDidFinishLaunching(_ notification: Notification) {
-        menubarItem = MenubarItem(title: "Setup Keyrace")
-
-        gh = GitHub()
-        menubarItem?.gh = gh
-
-        keyTap = KeyTap(self)
-        menubarItem?.keyTap = keyTap
-        keyTap!.getAccessibilityPermissions()
-    }
-    
-    func applicationWillTerminate(_ aNotification: Notification) {
-        // Insert code here to tear down your application
-    }
-}
-
-extension NSImage {
-    // Copies this image to a new one with a circular mask.
-    func circle() -> NSImage {
-        let image = NSImage(size: size)
-        image.lockFocus()
-
-        NSGraphicsContext.current?.imageInterpolation = .high
-        let frame = NSRect(origin: .zero, size: size)
-        NSBezierPath(ovalIn: frame).addClip()
-        draw(at: .zero, from: frame, operation: .sourceOver, fraction: 1)
-
-        image.unlockFocus()
-        return image
     }
 }
